+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 See the [upgrade guide](https://docs.pola.rs/releases/upgrade/0.20/)
 
 - Updated Polars to 0.38.1
+- Changed `count` method to exclude null values
 - Changed `dtype` and `schema` methods to always return instances of data types
 - Added `Enum` type
 - Added `sink_ipc`, `sink_csv`, and `sink_ndjson` methods to `LazyFrame`